JENKINS-5347 Fixed - Added use commit times on files
This fixes JENKINS-5347, we have a third party tool that uses the
timestamp on files to notify a cache in the tool to be updated. This fix
will add an option to the Subversion configuration to use commit times.
As the "use-commit-times" is a global configuration but we are only
interested in it as a per-project configuration to not affect our other
builds.