[FIXED JENKINS-18574]
AffectedFile contract specifies "path in the workspace that was affected".

Until now, Subversion plugin returned the full path relative to repo
root instead.

This commit adds storing the relative path to a file in the workspace
to the changelog.xml, and returns that path for AffectedFile.getPath().

Existing Subversion-specific consumers need to switch to getValue()
if they're not using it already, which continues to return the full
path within the repository (e.g. repository browsers).