Improve svn revert with as-it-is setting
When unknown is passed to the SvnRevert operation of SVNKit it's coerced
to empty, which means nothing actually gets reverted. So instead of
passing unknown pass infinity for revert as the svn depth. This seems to
do what one would want i.e. revert what ever happens to be there but not
checkout any additional files.